backend/classifier/requirements.txt
fastapi
uvicorn
openai==0.28
python-dotenv
typing_extensions